pyproject.toml
[project]
name = "iotclima-backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "alex-candido", email = "alex.candido.tec@gmail.com"},
]

dependencies = [
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "django-extensions>=4.1",
    
    "django-allauth>=65.3.0",
    "dj-rest-auth>=7.0.1",
    "djangorestframework-simplejwt>=5.3.1",

    "psycopg2-binary>=2.9.10",
    "dj-database-url>=2.3.0",
    "python-dotenv>=1.0.1",
    "djangorestframework-gis>=1.0",

    "dependency-injector>=4.46.0",
    "pydantic-settings>=2.9.1",

    "markdown>=3.8",
    "colorama>=0.4.6",
    "faker>=37.1.0",
    "pydot>=4.0.0",
    "requests>=2.32.3",
    "geopy>=2.4.1",
    "pandas>=2.3.0"
]

requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = false

[tool.pdm.scripts]
manage="python ./src/manage.py"
server = "pdm run manage runserver 0.0.0.0:3333"
module = "python ./scripts/cli.py"
